+++ ui/src/components/MyGames/index.js
@@ -1,2 +1,2 @@
-export Continue from './Continue'
+export MyGames from './MyGames'
 export GamesList from './GamesList'
