+++ ui/src/services/RouteService.js
@@ -18,8 +18,8 @@ class RouteService {
     books: () => '/app/books',
     booksnews: () => '/app/books/news',
     bookslibrary: () => '/app/books/library',
-    continue: () => '/app/continue',
-    continuelist: () => '/app/continue/list',
+    mygames: () => '/app/mygames',
+    mygameslist: () => '/app/mygames/list',
     playgame: gameId => `/app/play/${gameId}`,
     trialgame: bookId => `/app/trial/${bookId}`,
     write: () => '/app/write',
